requirements.txt
boto3
discord
python-dotenv
requests
pyjwt[crypto]
pynacl